Exported SII Table parser function
